tb4_desktop

+++ turtlebot4_desktop/update-rosdep.sh
@@ -2,9 +2,12 @@
 
 set -e
 
+rm -rf src
 mkdir -p src
-vcs import --input deps.repos src
-xargs -a unnecessary_ros2_pkgs.txt -I {} rm -rf src/{}
+vcs import --force --shallow --recursive --input deps.repos src
+if [[ -f "unnecessary_ros2_pkgs.txt" ]]; then
+    xargs -a unnecessary_ros2_pkgs.txt -I {} rm -rf src/{}
+fi
 
 source ../ros2_setup.sh
 
@@ -13,3 +16,8 @@ rosdep install \
     --from-paths src \
     --ignore-src \
     -s | awk '{print $5}' | sed -E '/^\s*$/d' | sort -n > rosdep.txt
+
+# sed -i \
+#     -e '/^cmake$/d' \
+#     -e '/^git$/d' \
+#     rosdep.txt
